third_party/blink/web_tests/editing/pasteboard/copy_paste_unselectable_content.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
selection_test(
    [
      '^1<span style="user-select: all;">2</span>3',
      '4<span style="user-select: none;">5</span>6|'
    ],
    selection => {
      assert_own_property(window, 'internals',
        'This test requires window.internals');
      assert_equals(selection.window.internals.selectedTextForClipboard(), '12346');
      assert_equals(selection.window.internals.selectedHTMLForClipboard()
        .replace(/ style=".+?"/g, ' style="..."'),
          '<span style=\"...\">1</span><span style=\"...\">2</span><span style=\"...\">34</span>' +
          '<span style=\"...\">6</span>');
    },
    [
      '^1<span style="user-select: all;">2</span>3',
      '4<span style="user-select: none;">5</span>6|'
    ],
    'Copy selection that contains basic unselectable content.');

selection_test(
    [
      '^<span style="user-select: none;">1<span>2</span></span>',
      '<span style="user-select: none;">3<span style="user-select: all;">4</span></span>',
      '<span style="user-select: none;">5<span>6</span><span style="user-select: all;">7</span>',
      '</span>|',
    ],
    selection => {
      assert_own_property(window, 'internals',
        'This test requires window.internals');
      assert_equals(selection.window.internals.selectedTextForClipboard(), '47');
      assert_equals(selection.window.internals.selectedHTMLForClipboard()
        .replace(/ style=".+?"/g, ' style="..."'),
        '<span style=\"...\">4</span><span style=\"...\"><span style=\"...\">7</span></span>');
    },
    [
      '^<span style="user-select: none;">1<span>2</span></span>',
      '<span style="user-select: none;">3<span style="user-select: all;">4</span></span>',
      '<span style="user-select: none;">5<span>6</span><span style="user-select: all;">7</span>',
      '</span>|',
    ],
    'Copy selection that contains nested unselectable and selectable content.');
</script>
